A graphical database diagram has been added. UserId is no longer transmitted via dto, but is clung from the authorization token. Added the ability to create a superuser.

+++ src/albums/albums.service.ts
@@ -1,20 +1,28 @@
-import {Injectable} from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from "@nestjs/common";
 import {Album} from "./albums.model";
 import {FilesService, FileType} from "../files/files.service";
 import {CreateAlbumDto} from "./dto/create-album.dto";
 import {InjectModel} from "@nestjs/sequelize";
 import { Song } from "../songs/songs.model";
+import { Musician } from "../musicians/musicians.model";
 
 @Injectable()
 export class AlbumsService {
     constructor(@InjectModel(Album) private albumRepository: typeof Album,
                 @InjectModel(Song) private songRepository: typeof Song,
+                @InjectModel(Musician) private musicianRepository: typeof Musician,
                 private fileService: FilesService) {}
 
 
-    async create(dto: CreateAlbumDto, image: any) {
+    async create(dto: CreateAlbumDto, image: any, userId: number): Promise<Album> {
+        const musician = await this.musicianRepository.findOne({
+          where: {userId: userId}
+        })
+        if (!musician) {
+          throw new HttpException("Музыкант не найден", HttpStatus.NOT_FOUND)
+        }
         const fileName = await this.fileService.createFile(image, FileType.ALBUMS_IMAGE)
-        const album = await this.albumRepository.create({...dto, image: fileName})
+        const album = await this.albumRepository.create({...dto, image: fileName, musicianId: musician.id})
         return album
     }
 
@@ -27,18 +35,35 @@ export class AlbumsService {
     return albums
   }
 
-  async deleteAlbumById(value: number): Promise<Album> {
+  async deleteAlbumById(value: number, userId: number): Promise<Album> {
       const album = await this.albumRepository.findByPk(value,{include:{all: true}});
-      await this.songRepository.destroy({
-        where: {
-          albumId: value,
-        }
-      })
-      await this.albumRepository.destroy({
-      where: {
-        id: value,
+      if (!album) {
+        throw new HttpException("Альбом не найден", HttpStatus.NOT_FOUND)
       }
-    })
-    return album;
+
+      const musician = await this.musicianRepository.findOne({
+        where: {userId: userId}
+      })
+      if (!musician) {
+        throw new HttpException("Исполнитель не найден", HttpStatus.NOT_FOUND)
+      }
+
+      if (album.musicianId === musician.id) {
+        await this.songRepository.destroy({
+          where: {
+            albumId: value,
+          }
+        })
+        await album.destroy();
+        return album;
+      }
+
+      throw new HttpException("Ошибка при удалении альбома", HttpStatus.BAD_REQUEST);
+
+      // await this.albumRepository.destroy({
+      // where: {
+      //   id: value,
+      // }
+      // })
   }
 }
